refactor: move modal to its own file
