+++ src/components/PostTimeline/index.js
@@ -2,10 +2,8 @@ import { useNavigate } from "react-router-dom";
 import { ReactTagify } from "react-tagify";
 import { FaTrashAlt, FaPencilAlt } from "react-icons/fa";
 import * as S from "./styles";
-import Modal from "react-modal";
 import { useState } from "react";
-
-Modal.setAppElement("#root")
+import DeleteModal from "../Modal";
 
 const PostTimeline = (props) => {
 	const [modalIsOpen, setModalIsOpen] = useState(false);
@@ -44,51 +42,13 @@ const PostTimeline = (props) => {
 					></S.ContainerMetadata>
 				</S.ContainerContent>
 			</S.ContainerPost>
-			<Modal
-				isOpen={modalIsOpen}
-				onRequestClose={() => setModalIsOpen(false)}
-				style={modalStyle}
-				contentLabel="Example Modal"
-			>
-				<S.ModalDiv>
-					<S.ModalText>
-						Are you sure you want to delete this post?
-					</S.ModalText>
-					<div>
-						<S.ModalButton
-							className="cancel"
-							onClick={() => setModalIsOpen(false)}
-						>
-							No, go back
-						</S.ModalButton>
-						<S.ModalButton
-							className="confirm"
-							onClick={() => setModalIsOpen(false)}
-						>
-							Yes, delete it
-						</S.ModalButton>
-					</div>
-				</S.ModalDiv>
-			</Modal>
+			<DeleteModal
+				id={props.id}
+				modalIsOpen={modalIsOpen}
+				setModalIsOpen={setModalIsOpen}
+			/>
 		</>
 	);
-};
-
-const modalStyle = {
-	content: {
-		top: "50%",
-		left: "50%",
-		right: "auto",
-		bottom: "auto",
-		marginRight: "-50%",
-		transform: "translate(-50%, -50%)",
-		backgroundColor: "#333333",
-		borderRadius: "50px 130px",
-		padding: "40px",
-		display: "flex",
-		justifyContent: "center",
-		alignItems: "center",
-	},
 };
 
 export default PostTimeline;
